+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 
 ### Added
 - Added global variable for a printing template.
+- \_remove_files:  Private function for process to decrease complexity.
 - \_process_item:  Private function for process to decrease complexity.
 
 ### Changed